#93 Latest version of VisualHFSM. It includes the following improvements:
Better system of ICE interfaces additioning. Now, the user will have to enter
a name (of the server interface), an IP, a port and an interface type (Camera,
Pose3DMotors, ...). The types are dynamically generated from the files under
/usr/local/include/jderobot/slice

Now, before to generate the C++ code, it checks:
- No repeated ICE interfaces.
- No empty state names.
- No repeated state names.
- No empty transition types.
- No empty iteration time for a subautomata.
- No empty subautomatas.

When changing a state name, the treeview updates its information.

There is now the chance of add headers (for external libraries).

Fixed a bug in copying a node.

There are no old files (of the oldest version of VisualHFSM).